+++ Panels/Header.js
@@ -18,6 +18,7 @@ import $L from '../internal/$L';
 import Button from '../Button';
 import Heading from '../Heading';
 import Skinnable from '../Skinnable';
+import {useScrollPosition} from '../useScroll/useScrollPosition';
 import WindowEventable from '../internal/WindowEventable';
 
 import {PanelsStateContext} from './Viewport';
@@ -97,11 +98,9 @@ const HeaderBase = kind({
 		backButtonAvailable: PropTypes.bool,
 
 		/**
-		 * The background opacity of the application back button.
+		 * Background opacity of the application back button.
 		 *
-		 * * Values: `'opaque'`, `'transparent'`
-		 *
-		 * @type {String}
+		 * @type {('opaque'|'transparent')}
 		 * @default 'transparent'
 		 * @public
 		 */
@@ -131,7 +130,7 @@ const HeaderBase = kind({
 		]),
 
 		/**
-		 * Sets the hint string read when focusing the application close button.
+		 * Hint string read when focusing the application close button.
 		 *
 		 * @type {String}
 		 * @default 'Exit app'
@@ -140,11 +139,9 @@ const HeaderBase = kind({
 		closeButtonAriaLabel: PropTypes.string,
 
 		/**
-		 * The background opacity of the application close button.
+		 * Background opacity of the application close button.
 		 *
-		 * * Values: `'opaque'`, `'transparent'`
-		 *
-		 * @type {String}
+		 * @type {('opaque'|'transparent')}
 		 * @default 'transparent'
 		 * @public
 		 */
@@ -175,6 +172,19 @@ const HeaderBase = kind({
 		 * @private
 		 */
 		entering: PropTypes.bool,
+
+		/**
+		 * Minimizes the Header to only show the header components in order to feature the panel
+		 * content more prominately.
+		 *
+		 * Has no effect on `type="compact"`. When a `Header` is used inside a
+		 * [`Panel`]{@link sandstone/Panels.Panel} with `featureContent` set it will automatically
+		 * collapse unless overridden by this prop.
+		 *
+		 * @type {Boolean}
+		 * @private
+		 */
+		featureContent: PropTypes.bool,
 
 		/**
 		 * [`Input`]{@link sandstone/Input} element that will replace the `title`.
@@ -372,6 +382,18 @@ const HeaderBase = kind({
 		]),
 
 		/**
+		 * The method which receives the reference node to the title element, used to determine
+		 * the `titleMeasurements`.
+		 *
+		 * @type {Function|Object}
+		 * @private
+		 */
+		titleRef: PropTypes.oneOfType([
+			PropTypes.func,
+			PropTypes.shape({current: PropTypes.any})
+		]),
+
+		/**
 		 * Set the type of header to be used.
 		 *
 		 * @type {('compact'|'dense'|'mini'|'standard')}
@@ -418,8 +440,9 @@ const HeaderBase = kind({
 	computed: {
 		backButtonAriaLabel: preferPropOverContext('backButtonAriaLabel'),
 		backButtonBackgroundOpacity: preferPropOverContext('backButtonBackgroundOpacity'),
-		className: ({backButtonAvailable, hover, noBackButton, entering, centered, children, slotAbove, type, styler}) => styler.append(
+		className: ({backButtonAvailable, featureContent, hover, noBackButton, entering, centered, children, slotAbove, type, styler}) => styler.append(
 			{
+				featureContent,
 				centered,
 				// This likely doesn't need to be as verbose as it is, with the first 2 conditionals
 				showBack: (backButtonAvailable && !noBackButton && (hover || entering)),
@@ -462,9 +485,11 @@ const HeaderBase = kind({
 		slotBefore,
 		subtitle,
 		title,
+		titleRef,
 		type,
 		...rest
 	}) => {
+		delete rest.featureContent;
 		delete rest.entering;
 		delete rest.onHideBack;
 		delete rest.onShowBack;
@@ -543,7 +568,7 @@ const HeaderBase = kind({
 		return (
 			<header {...rest}>
 				{slotAbove ? <nav className={css.slotAbove}>{slotAbove}</nav> : null}
-				<Row className={css.titlesRow} align="center">
+				<Row className={css.titlesRow} align="center" ref={titleRef}>
 					{(bothBeforeAndAfter || slotBefore || backButton) ? (
 						<Cell className={css.slotBefore} shrink={!bothBeforeAndAfter}>{backButton}{slotBefore}</Cell>
 					) : null}
@@ -571,9 +596,17 @@ const HeaderBase = kind({
 	}
 });
 
+const CollapsingHeaderDecorator = (Wrapped) => {
+	return function CollapsingHeaderDecorator (props) { // eslint-disable-line no-shadow
+		const {shouldFeatureContent} = useScrollPosition() || {};
+		return <Wrapped featureContent={shouldFeatureContent} {...props} />;
+	};
+};
+
 const HeaderDecorator = compose(
 	Slottable({slots: ['headerInput', 'title', 'subtitle', 'slotAbove', 'slotAfter', 'slotBefore']}),
 	Skinnable,
+	CollapsingHeaderDecorator,
 	Toggleable({prop: 'hover', activate: 'onShowBack', deactivate: 'onHideBack', toggle: null}),
 	WindowEventable({globalNode: 'document', onKeyDown: handleWindowKeyPress})
 );