+++ Panels/WizardPanel.js
@@ -84,6 +84,14 @@ const WizardPanelBase = kind({
 		nextButtonText: PropTypes.string,
 
 		/**
+		 * Disables panel transitions.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		noAnimation: PropTypes.bool,
+
+		/**
 		* Called when the index value is changed.
 		*
 		* @type {Function}
@@ -167,7 +175,7 @@ const WizardPanelBase = kind({
 		}
 	},
 
-	render: ({buttons, children, footer, index, total, nextButtonText, onIncrementStep, onDecrementStep, prevButtonText, reverseTransition, rtl, subtitle, title, ...rest}) => {
+	render: ({buttons, children, footer, index, total, nextButtonText, noAnimation, onIncrementStep, onDecrementStep, prevButtonText, reverseTransition, rtl, subtitle, title, ...rest}) => {
 		return (
 			<Panel {...rest}>
 				<Header
@@ -211,6 +219,7 @@ const WizardPanelBase = kind({
 							<ViewManager
 								arranger={SlideLeftArranger}
 								duration={400}
+								noAnimation={noAnimation}
 								reverseTransition={reverseTransition}
 							>
 								{children}