+++ TabLayout/TabLayout.js
@@ -3,12 +3,14 @@
  *
  * @module sandstone/TabLayout
  * @exports TabLayout
+ * @exports Tab
  */
 
 import {adaptEvent, forward, handle} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import {Changeable} from '@enact/ui/Changeable';
 import {Cell, Layout} from '@enact/ui/Layout';
+import deprecate from '@enact/core/internal/deprecate';
 import Toggleable from '@enact/ui/Toggleable';
 import ViewManager from '@enact/ui/ViewManager';
 import PropTypes from 'prop-types';
@@ -16,11 +18,25 @@ import compose from 'ramda/src/compose';
 import React from 'react';
 
 import TabGroup from './TabGroup';
+import Tab from './Tab';
 
 import componentCss from './TabLayout.module.less';
 
 /**
  * Tabbed Layout component.
+ *
+ * Example:
+ *
+ * ```jsx
+ * 	<TabLayout>
+ * 		<Tab title="Tab One">
+ * 			<Item>Hello</Item>
+ * 		</Tab>
+ * 		<Tab title="Tab Two">
+ * 			<Item>Goodbye</Item>
+ * 		</Tab>
+ * 	</TabLayout>
+ * ```
  *
  * @class TabLayout
  * @memberof sandstone/TabLayout
@@ -32,27 +48,12 @@ const TabLayoutBase = kind({
 
 	propTypes: /** @lends sandstone/TabLayout.TabLayout.prototype */ {
 		/**
-		 * List of tabs to display.
+		 * Collection of [Tabs]{@link sandstone/TabLayout.Tab} to render.
 		 *
-		 * Each object in the array of tabs should include a `title` property and, optionally, an
-		 * `icon` property (see: {@link sandstone/Icon.IconBase.children}). If an icon is not
-		 * supplied for any tabs, no icons will be displayed when collapsed.
-		 *
-		 * @type {Object[]}
-		 * @required
+		 * @type {Node}
 		 * @public
 		 */
-		tabs: PropTypes.array.isRequired,
-
-		/**
-		 * List of content to be rendered for each tab.
-		 *
-		 * The number of children should match the number of tabs.
-		 *
-		 * @type {Node[]}
-		 * @public
-		 */
-		children: PropTypes.arrayOf(PropTypes.node),
+		children: PropTypes.node,
 
 		/**
 		 * Collapses the vertical tab list into icons only.
@@ -118,7 +119,21 @@ const TabLayoutBase = kind({
 		 * @default 'vertical'
 		 * @public
 		 */
-		orientation: PropTypes.oneOf(['horizontal', 'vertical'])
+		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
+
+		/**
+		 * List of tabs to display.
+		 *
+		 * Each object in the array of tabs should include a `title` property and, optionally, an
+		 * `icon` property (see: {@link sandstone/Icon.IconBase.children}). If an icon is not
+		 * supplied for any tabs, no icons will be displayed when collapsed.
+		 *
+		 * @type {Object[]}
+		 * @deprecated To be removed in 1.0.0-beta.1. Use
+		 *	[Tab.title]{@link sandstone/TabLayout.Tab.title} instead.
+		 * @public
+		 */
+		tabs: PropTypes.array
 	},
 
 	defaultProps: {
@@ -138,14 +153,31 @@ const TabLayoutBase = kind({
 	},
 
 	computed: {
+		children: ({children, tabs}) => {
+			if (tabs) {
+				deprecate({
+					name: 'sandstone/TabLayout.TabLayout.tabs',
+					until: '1.0.0-beta.1'
+				});
+				return children;
+			} else {
+				return React.Children.map(children, (child) => {
+					return <React.Fragment>{child.props.children}</React.Fragment>;
+				});
+			}
+		},
 		className: ({collapsed, orientation, styler}) => styler.append(
 			{collapsed: orientation === 'vertical' && collapsed},
 			orientation
 		),
 		tabOrientation: ({orientation}) => orientation === 'vertical' ? 'horizontal' : 'vertical',
 		// limit to 5 tabs for horizontal orientation
-		tabs: ({orientation, tabs}) => {
-			return orientation === 'horizontal' && tabs.length > 5 ? tabs.slice(0, 5) : tabs;
+		tabs: ({children, orientation, tabs}) => {
+			let outTabs = tabs || React.Children.map(children, (child) => {
+				const {icon, title} = child.props;
+				return {icon, title};
+			});
+			return orientation === 'horizontal' && outTabs.length > 5 ? outTabs.slice(0, 5) : outTabs;
 		}
 	},
 
@@ -187,8 +219,18 @@ const TabLayoutDecorator = compose(
 // Currently not documenting the base output since it's not exported
 const TabLayout = TabLayoutDecorator(TabLayoutBase);
 
+/**
+ * A shortcut to access {@link sandstone/TabLayout.Tab}
+ *
+ * @name Tab
+ * @static
+ * @memberof sandstone/TabLayout.TabLayout
+ */
+TabLayout.Tab = Tab;
+
 export default TabLayout;
 export {
 	TabLayout,
-	TabLayoutBase
+	TabLayoutBase,
+	Tab
 };